adopted new dropdown menu for combobox
