+++ customtkinter/assets/themes/green.json
@@ -31,9 +31,9 @@
     "optionmenu_button_hover":["gray40", "gray70"],
     "combobox_border": ["gray70", "gray32"],
     "combobox_button_hover": ["#6E7174", "#7A848D"],
-    "dropdown_color": ["#A8ACB1", "#535353"],
-    "dropdown_hover": ["#D6DCE2", "#393D40"],
-    "dropdown_text": ["gray12", "gray90"]
+    "dropdown_color": ["gray90", "gray20"],
+    "dropdown_hover": ["gray75", "gray28"],
+    "dropdown_text": ["gray10", "#DCE4EE"]
   },
   "text": {
     "macOS": {
