+++ test/manual_integration_tests/test_variables.py
